+++ packages/rs-dpp/src/data_contract/group/mod.rs
@@ -1,6 +1,7 @@
 use crate::data_contract::group::accessors::v0::{GroupV0Getters, GroupV0Setters};
 use crate::data_contract::group::methods::v0::GroupMethodsV0;
 use crate::data_contract::group::v0::GroupV0;
+use crate::data_contract::GroupContractPosition;
 use crate::errors::ProtocolError;
 use crate::validation::SimpleConsensusValidationResult;
 use bincode::{Decode, Encode};
@@ -92,10 +93,11 @@ impl GroupV0Setters for Group {
 impl GroupMethodsV0 for Group {
     fn validate(
         &self,
+        group_contract_position: Option<GroupContractPosition>,
         platform_version: &PlatformVersion,
     ) -> Result<SimpleConsensusValidationResult, ProtocolError> {
         match self {
-            Group::V0(group_v0) => group_v0.validate(platform_version),
+            Group::V0(group_v0) => group_v0.validate(group_contract_position, platform_version),
         }
     }
 }