+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/data_contract_update/basic_structure/v0/mod.rs
@@ -3,9 +3,11 @@ use dpp::consensus::basic::data_contract::{
     InvalidTokenBaseSupplyError, NewTokensDestinationIdentityOptionRequiredError,
     NonContiguousContractTokenPositionsError,
 };
+use dpp::dashcore::Network;
 use dpp::data_contract::associated_token::token_configuration::accessors::v0::TokenConfigurationV0Getters;
 use dpp::data_contract::associated_token::token_distribution_rules::accessors::v0::TokenDistributionRulesV0Getters;
 use dpp::data_contract::associated_token::token_perpetual_distribution::methods::v0::TokenPerpetualDistributionV0Accessors;
+use dpp::data_contract::change_control_rules::authorized_action_takers::AuthorizedActionTakers;
 use dpp::data_contract::TokenContractPosition;
 use dpp::prelude::DataContract;
 use dpp::state_transition::data_contract_update_transition::accessors::DataContractUpdateTransitionAccessorsV0;
@@ -17,6 +19,7 @@ pub(in crate::execution::validation::state_transition::state_transitions::data_c
 {
     fn validate_basic_structure_v0(
         &self,
+        network_type: Network,
         platform_version: &PlatformVersion,
     ) -> Result<SimpleConsensusValidationResult, Error>;
 }
@@ -24,6 +27,7 @@ pub(in crate::execution::validation::state_transition::state_transitions::data_c
 impl DataContractUpdateStateTransitionBasicStructureValidationV0 for DataContractUpdateTransition {
     fn validate_basic_structure_v0(
         &self,
+        network_type: Network,
         platform_version: &PlatformVersion,
     ) -> Result<SimpleConsensusValidationResult, Error> {
         let groups = self.data_contract().groups();
@@ -73,6 +77,16 @@ impl DataContractUpdateStateTransitionBasicStructureValidationV0 for DataContrac
                 .distribution_rules()
                 .perpetual_distribution()
             {
+                // we validate the interval (that it's more than one hour or over 100 blocks)
+                // also that if it is time based we are using minute intervals
+                let validation_result = perpetual_distribution
+                    .distribution_type()
+                    .validate_structure_interval(network_type, platform_version)?;
+
+                if !validation_result.is_valid() {
+                    return Ok(validation_result);
+                }
+
                 // We use 0 as the start moment to show that we are starting now with no offset
                 let validation_result = perpetual_distribution
                     .distribution_type()
@@ -91,6 +105,16 @@ impl DataContractUpdateStateTransitionBasicStructureValidationV0 for DataContrac
                 && !token_configuration
                     .distribution_rules()
                     .minting_allow_choosing_destination()
+                && !(token_configuration
+                    .distribution_rules()
+                    .minting_allow_choosing_destination_rules()
+                    .authorized_to_make_change_action_takers()
+                    == &AuthorizedActionTakers::NoOne
+                    && token_configuration
+                        .distribution_rules()
+                        .minting_allow_choosing_destination_rules()
+                        .admin_action_takers()
+                        == &AuthorizedActionTakers::NoOne)
             {
                 return Ok(SimpleConsensusValidationResult::new_with_error(
                     NewTokensDestinationIdentityOptionRequiredError::new(