+++ packages/rs-drive/src/drive/credit_pools/storage_fee_distribution_pool/get_storage_fees_from_distribution_pool/v0/mod.rs
@@ -52,7 +52,7 @@ mod tests {
 
         #[test]
         fn test_error_if_epoch_is_not_initiated() {
-            let drive = setup_drive(None, None);
+            let drive = setup_drive(None);
             let transaction = drive.grove.start_transaction();
 
             let platform_version = PlatformVersion::first();