+++ packages/rs-drive/src/drive/identity/contract_info/identity_contract_nonce/merge_identity_contract_nonce/v0/mod.rs
@@ -304,7 +304,7 @@ mod tests {
     use platform_version::version::PlatformVersion;
 
     fn setup_base_test(contract_id: [u8; 32]) -> (Drive, Identity) {
-        let drive = setup_drive(None, None);
+        let drive = setup_drive(None);
         let transaction = drive.grove.start_transaction();
 
         let platform_version = PlatformVersion::first();